requirements.txt
# Dependencies are as follows
pillow
